+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - ♿️(menu) add opensInNewWindow for external link aria-label #246
 - 🚸(LaGaufreV2) automatic positionning #260
+- ✨(filter) add custom sub-element support to options
+- ✨(help-menu) add legal links and custom options
 
 ## 0.23.2